TBM3 - update for user location
For Edinburgh graany can only be in 1 of 2 places. We will be told up front and know the locations in UR coord system.  The places will be sent as the judges coord system and the code will now remap based on the judges coords being in a range of x and y values. Hence we can equste to our coord system.